--- a/src/lib.rs
+++ b/src/lib.rs
@@ -41,18 +41,27 @@ extern crate std;
 
 extern crate alloc;
 
+macro_rules! reverse_apply {
+    ($m: ident [] $($reversed:tt)*) => {
+        $m!{$($reversed),*}  // base case
+    };
+    ($m: ident [$first:tt $($rest:tt)*] $($reversed:tt)*) => {
+        reverse_apply!{$m [$($rest)*] $first $($reversed)*}
+    };
+}
+
 /// Imagine macro parameters, but more like those Russian dolls.
 ///
-/// Calls m!(A, B, C), m!(A, B), m!(B), and m!() for i.e. (m, A, B, C)
+/// Calls m!(), m!(A), m!(A, B), and m!(A, B, C) for i.e. (m, A, B, C)
 /// where m is any macro, for any number of parameters.
 macro_rules! smaller_tuples_too {
-    ($m: ident, $ty: ident) => {
+    ($m: ident, $next: tt) => {
         $m!{}
-        $m!{$ty}
+        $m!{$next}
     };
-    ($m: ident, $ty: ident, $($tt: ident),*) => {
-        smaller_tuples_too!{$m, $($tt),*}
-        $m!{$ty, $($tt),*}
+    ($m: ident, $next: tt, $($rest: tt),*) => {
+        smaller_tuples_too!{$m, $($rest),*}
+        reverse_apply!{$m [$next $($rest)*]}
     };
 }