@Ralith 2023-09-16
Oddly this was much less impactful on my Ryzen 7800x3d desktop than my old i7-8550U laptop, though the outlining of `next_archetype` fixed `iterate_mut` being significantly slower than `iterate` there.